Refactored Audio player
