+++ Frontend/src/components/index.ts
@@ -1,3 +1,4 @@
 export { default as Tesseract } from './Tesseract'
 export { default as Side } from './Side'
 export { default as IconButton } from './IconButton'
+export { default as AudioPlayer } from './AudioPlayer'
